refactor: Migrate orbit-tools tests to per-module tests/ layout [ORB-00232]
