+++ crates/orbit-tools/src/builtin/git/tests/mod.rs
@@ -1,0 +1,3 @@
+#![allow(missing_docs)]
+
+mod relative_paths;
