+++ crates/orbit-tools/src/builtin/github/pr_comment_reply/tests/mod.rs
@@ -1,0 +1,3 @@
+#![allow(missing_docs)]
+
+mod response_parsing;
