+++ crates/orbit-tools/src/builtin/orbit/friction/add.rs
@@ -56,31 +56,4 @@ impl Tool for OrbitFrictionAddTool {
 }
 
 #[cfg(test)]
-mod tests {
-    use orbit_common::friction::{DEFAULT_FRICTION_TAGS, friction_tags_literal};
-
-    use super::*;
-
-    #[test]
-    fn tags_parameter_description_lists_default_taxonomy() {
-        let schema = OrbitFrictionAddTool.schema();
-        let tags_param = schema
-            .parameters
-            .iter()
-            .find(|param| param.name == "tags")
-            .expect("tags parameter");
-
-        assert!(
-            tags_param.description.contains(&friction_tags_literal()),
-            "{}",
-            tags_param.description
-        );
-        for (tag, _description) in DEFAULT_FRICTION_TAGS {
-            assert!(
-                tags_param.description.contains(tag),
-                "tags description should include {tag}: {}",
-                tags_param.description
-            );
-        }
-    }
-}
+mod tests;
