+++ crates/orbit-tools/src/builtin/orbit/knowledge/pack.rs
@@ -202,22 +202,4 @@ fn summarize_pack_entry(entry: &mut Value) {
 }
 
 #[cfg(test)]
-mod tests {
-    use serde_json::json;
-
-    use super::add_refresh_diagnostics;
-
-    #[test]
-    fn refresh_diagnostics_only_report_actual_auto_refresh_skips() {
-        let mut refreshed_pack = json!({"entries": []});
-        add_refresh_diagnostics(&mut refreshed_pack, false, None, false);
-        assert!(refreshed_pack.get("diagnostics").is_none());
-
-        let mut skipped_pack = json!({"entries": []});
-        add_refresh_diagnostics(&mut skipped_pack, true, None, false);
-        assert_eq!(
-            skipped_pack["diagnostics"]["auto_refresh"]["status"],
-            "skipped"
-        );
-    }
-}
+mod tests;
