+++ crates/orbit-tools/src/builtin/orbit/knowledge/pack/tests/mod.rs
@@ -1,0 +1,3 @@
+#![allow(missing_docs)]
+
+mod diagnostics;
