+++ crates/orbit-tools/src/builtin/orbit/task/artifact_put/tests/mod.rs
@@ -1,0 +1,3 @@
+#![allow(missing_docs)]
+
+mod execution;
