+++ crates/orbit-tools/src/builtin/orbit/task/update/tests/mod.rs
@@ -1,0 +1,3 @@
+#![allow(missing_docs)]
+
+mod source_task_id;
