+++ crates/orbit-tools/src/builtin/orbit/tests/mod.rs
@@ -1,0 +1,3 @@
+#![allow(missing_docs)]
+
+mod identity;
